test(red) checks home_page title using url route
Originally it would be written using assertContains from
the django test suit, but I couldn't put the pycharm equivalent
to work. Because of that, I went with the .content.decode() route,
which work fairly well.

the reason I couldn't make the pycharm equivalent work (the
equivalent being ' from pytest_django.asserts import assertContains')
is because when I try to assertContains to check for the html of
the request, it checks inside the:
  '<HttpResponse status_code=200, "text/html; charset=utf-8">'
when it should be checking for the content of the response.