feat(wip) add ability to make form post  on home view
wip because it's not really what I want when I mean
'ability to make form post', the list is hardcoded and
the logic is not there yet, but we can, technically, make
a post from the form.

the second assert is there just to remember(me) that
the the list is hardcoded, and even though the test passed,
it's because of cheating. It will be refactored.